+++ src/ui/views/ItemCraftView.cpp
@@ -43,7 +43,7 @@ ItemCraftView::ItemCraftView(ViewManager* gvm) : View(gvm), school(NATURE), curr
     updateItemName();
   });
   
-  for (int i = 0; i < sizeof(ITEM_TYPES)/sizeof(ITEM_TYPES[0]); ++i)
+  for (size_t i = 0; i < sizeof(ITEM_TYPES)/sizeof(ITEM_TYPES[0]); ++i)
   {
     RadioButton<Item::TypeID>* button = RadioButton<Item::TypeID>::build("type", ITEM_TYPES[i], itemType, 156 + 33*(i%5), 3 + 15*(i/5), TextureID::ITEM_CRAFT_BUTTONS, i, 0, 2);
     buttons.push_back(button);
